+++ libs/tools/generators/json-schema/src/generator/generator-config.ts
